Add deferred scalar counts; condense source comments

+++ gibbon-compiler/examples/soa_examples/programs/AOS/LinearListReduction.hs
@@ -1,5 +1,6 @@
--- @BENCH adt_fields=11
-data List = Cons Int Int Int Int Int Int Int Int Int Int List | Nil
+-- LinearListReduction: List (Linear).
+-- Functions: mkList, reduce.
+data List = Cons Int64 Int64 Int64 Int64 Int64 Int64 Int64 Int64 Int64 Int64 List | Nil
 
 {-# ANN type List "Linear" #-}
