+++ gibbon-compiler/examples/soa_examples/programs/AOS/reduceNestedList.hs
@@ -1,7 +1,7 @@
--- @BENCH adt_fields=3
--- @BENCH adt_type=List
-data ListA = ConsA Int ListA | NilA
-data List = Cons Int ListA List | Nil
+-- reduceNestedList: ListA (Linear), List (Linear).
+-- Functions: mkListA, mkList, reduce.
+data ListA = ConsA Int64 ListA | NilA
+data List = Cons Int64 ListA List | Nil
 
 {-# ANN type ListA "Linear" #-}
 {-# ANN type List "Linear" #-}
